add some func and fix some errors

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: efumiko <efumiko@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/12 13:04:21 by efumiko           #+#    #+#             */
-/*   Updated: 2020/05/12 14:44:58 by efumiko          ###   ########.fr       */
+/*   Updated: 2020/05/23 11:23:29 by efumiko          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	long long int i;
+	int i;
 
 	i = 0;
 	while (s[i] != 0)
